feat: add a signIn and signUp component
